+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - [K8s] Added configuration for pod and container `securityContext`.
 - [Docs] Corrected MinIO/Ceph config template keys and removed obsolete Kubernetes image references.
 - [GCP Functions] Updated `gcp_functions` backend to Google Cloud Run functions (Cloud Functions v2 API).
+- [Python] Updated all backend python versions (deprecated Python 3.9, added Python 3.14)
+- [Azure Funcions] Updated default funcions plan to Flex Consumption
 
 ### Fixed
 - [K8s] Fixed default runtime builds impacted by Debian Buster end-of-life.